# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Official documentation to consume the API.",
        "title": "College-app API",
        "contact": {
            "name": "Jesús 'VoN' Rodríguez",
            "url": "https://www.linkedin.com/in/jes%C3%BAs-%C3%A1ngel-rodr%C3%ADguez-mart%C3%ADnez-84991a1b4/",
            "email": "sudo.von.contact@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/advices": {
            "get": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "List advices.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advices"
                ],
                "summary": "List advices.",
                "operationId": "list-advices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID that will attend to the advice.",
                        "name": "will_attend",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "User ID that will teach the advice.",
                        "name": "will_teach",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.AdviceList"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Create advice.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advices"
                ],
                "summary": "Create advice.",
                "operationId": "create-advice",
                "parameters": [
                    {
                        "description": "Advice that wants to be stored.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.AdvicePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/advices/{id}": {
            "get": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Get advice given its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advices"
                ],
                "summary": "Show advice.",
                "operationId": "get-advice-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Advice ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.AdviceResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Delete advice given its ID.",
                "tags": [
                    "advices"
                ],
                "summary": "Delete advice.",
                "operationId": "delete-advice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Advice ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Update advice given its ID.",
                "tags": [
                    "advices"
                ],
                "summary": "Update advice.",
                "operationId": "update-advice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Advice ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Advice information that wants to be updated.",
                        "name": "updatePayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.UpdateAdvicePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/advices/{id}/students-number": {
            "patch": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Update the number of students who will attend the advice.",
                "tags": [
                    "advices"
                ],
                "summary": "Update advice students number.",
                "operationId": "update-advice-students-number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Advice ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/contacts/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Show contact given the user ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Show contact by user ID.",
                "operationId": "show-contact-by-user-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ContactResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Create contact for a specific user given its ID.",
                "tags": [
                    "contacts"
                ],
                "summary": "Create contact.",
                "operationId": "create-contact-by-user-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Contact that wants to be stored.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.ContactPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/contacts/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Update contact given its ID.",
                "tags": [
                    "contacts"
                ],
                "summary": "Update contact.",
                "operationId": "update-contact-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contact ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Contact information that wants to be updated.",
                        "name": "updatePayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.UpdateContactPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Show basic user information.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Show user.",
                "operationId": "get-tiny-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.TinyUserResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Update user information.",
                "tags": [
                    "users"
                ],
                "summary": "Update user.",
                "operationId": "update-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User information that wants to be updated.",
                        "name": "updatePayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.UpdateUserPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "presenter.AdviceList": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.AdviceResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "presenter.AdvicePayload": {
            "type": "object",
            "properties": {
                "advice_date": {
                    "type": "string",
                    "example": "2021-10-24 15:04"
                },
                "classroom_id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622be"
                },
                "subject": {
                    "type": "string",
                    "example": "Chemistry"
                }
            }
        },
        "presenter.AdviceResponse": {
            "type": "object",
            "properties": {
                "advice_date": {
                    "type": "string",
                    "example": "2021-10-24 15:04"
                },
                "classroom": {
                    "$ref": "#/definitions/presenter.ClassroomResponse"
                },
                "id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622be"
                },
                "students_will_attend": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subject": {
                    "type": "string",
                    "example": "Chemistry"
                },
                "user": {
                    "$ref": "#/definitions/presenter.TinyUserResponse"
                }
            }
        },
        "presenter.ClassroomResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622be"
                },
                "name": {
                    "type": "string",
                    "example": "VoN"
                }
            }
        },
        "presenter.ContactPayload": {
            "type": "object",
            "properties": {
                "contact_name": {
                    "type": "string",
                    "example": "VoN"
                },
                "contact_number": {
                    "type": "string",
                    "example": "528661234567"
                },
                "message": {
                    "type": "string",
                    "example": "Hi, i need..."
                }
            }
        },
        "presenter.ContactResponse": {
            "type": "object",
            "properties": {
                "contact_name": {
                    "type": "string",
                    "example": "VoN"
                },
                "contact_number": {
                    "type": "string",
                    "example": "528661234567"
                },
                "id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622be"
                },
                "message": {
                    "type": "string",
                    "example": "Hi, i need..."
                },
                "user_id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622bd"
                }
            }
        },
...
        "presenter.TinyUserResponse": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "1997-04-17"
                },
                "email": {
                    "type": "string",
                    "example": "sudo.von.contact@gmail.com"
                },
                "id": {
                    "type": "string",
                    "example": "613aab4d8a6ef50007e622bd"
                },
                "name": {
                    "type": "string",
                    "example": "Sudo Von"
                },
                "registration_number": {
                    "type": "string",
                    "example": "16190770"
                }
            }
        },
...
        "presenter.UpdateAdvicePayload": {
            "type": "object",
            "properties": {
                "advice_date": {
                    "type": "string",
                    "example": "1997-04-17 15:04"
                },
                "classroom_id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622be"
                },
                "subject": {
                    "type": "string",
                    "example": "Chemistry"
                }
            }
        },
        "presenter.UpdateContactPayload": {
            "type": "object",
            "properties": {
                "contact_name": {
                    "type": "string",
                    "example": "VoN"
                },
                "contact_number": {
                    "type": "string",
                    "example": "528661234567"
                },
                "message": {
                    "type": "string",
                    "example": "Hi, i need..."
                }
            }
        },
        "presenter.UpdateUserPayload": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "1997-04-17"
                },
                "email": {
                    "type": "string",
                    "example": "sudo.von.contact@gmail.com"
                },
                "name": {
                    "type": "string",
                    "example": "Sudo Von"
                },
                "registration_number": {
                    "type": "string",
                    "example": "16190770"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerJWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}